refactor(agents): extract AgentLoaderDelegate from AgentEditorViewModel
Move the editor's initial fetches into AgentLoaderDelegate: the edited agent
(via the /expanded endpoint) plus the reference data (tool catalog, categories,
models, MCP tools, handoff/chain agent list). The VM init now calls
loadReferenceData() + loadAgent(); the delegate triggers the file-metadata
re-merge after applying agent data.

No behavior change.
